+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-* Ref: Deprecate cacheDirSize and add maxCacheItems
+* Ref: Deprecate cacheDirSize and add maxCacheItems (#1499)
 
 # 5.0.0-beta.6