+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Ref: Deprecate cacheDirSize and add maxCacheItems (#1499)
+
 # 5.0.0-beta.6
 
 * Feat: Add secondary constructor to SentryOkHttpInterceptor (#1491)