+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Ref: Deprecate SentryBaseEvent#getOriginThrowable and add SentryBaseEvent#getThrowableMechanism (#1502)
+* Fix: Graceful Shutdown flushes event instead of Closing SDK (#1500)
 * Fix: Do not append threads that come from the EnvelopeFileObserver (#1501)
 * Ref: Deprecate cacheDirSize and add maxCacheItems (#1499)